pyproject.toml
[project]
name = "autoanki"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.40",
    "genanki>=0.13",
    "openai>=1.0",
    "pymupdf>=1.23",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.scripts]
autoanki = "autoanki.__main__:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
